新增邮件发布工具类

+++ XG_Project_OC/XG_Project_OC/Utils/Category/NSString+XGAdd.m
@@ -236,7 +236,7 @@
 }
 
 /// 将目标时间戳转成当前时区的时间戳
-+ (NSDate *)getYearMonthDay:(NSInteger)timeStamp {
++ (NSDate *)getYearMonthDay:(long long)timeStamp {
     if (timeStamp <= 0) {
         return nil;
     }
